Combined src/clm/python/datasets.py and src/clm/datasets.py, removed src/clm/python.datasets.py and unused  codes from src/clm/datasets
