+++ src/clm/commands/inner_train_models_RNN.py
@@ -13,7 +13,7 @@ from tqdm import tqdm
 
 from rdkit import rdBase
 
-from clm.python.datasets import SmilesDataset, SelfiesDataset
+from clm.datasets import SmilesDataset, SelfiesDataset
 from clm.python.models import RNN
 from clm.python.loggers import EarlyStopping, track_loss, print_update
 from clm.functions import set_seed, seed_type, read_smiles, write_smiles
